+++ app/src/main/java/com/brandonjamesyoung/levelup/data/QuestHistoryRepository.kt
@@ -15,7 +15,7 @@ class QuestHistoryRepository @Inject constructor(
     fun observeAll() = questHistoryDao.observeAll()
 
     @WorkerThread
-    suspend fun insert(completedQuest: CompletedQuest) = externalScope.launch {
+    fun insert(completedQuest: CompletedQuest) = externalScope.launch {
         questHistoryDao.insert(completedQuest)
     }
 }